+++ src/core/api/auth/index.ts
@@ -1,7 +1,7 @@
 import type { IAuthApi } from './auth.interface'
-import { authApiReal } from './auth.http'
+import { authApiHttp } from './auth.http'
 import { authApiMock } from './auth.mock'
 
 const isMockEnabled = import.meta.env.VITE_USE_MOCK === 'true'
 
-export const authApi: IAuthApi = isMockEnabled ? authApiMock : authApiReal
+export const authApi: IAuthApi = isMockEnabled ? authApiMock : authApiHttp